requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Deprecated==1.2.13
Django==2.2.14
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==2.8
oauthlib==3.0.1
packaging==21.3
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.5.0
PyJWT==2.3.0
pyparsing==3.0.7
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
redis==4.1.4
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2
wrapt==1.13.3
